+++ crankshaft-engine/CHANGELOG.md
@@ -19,12 +19,18 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   new `resource-usage-metadata` configuration; when enabled, tasks are polled
   with the `BASIC` view ([#86](https://github.com/stjude-rust-labs/crankshaft/pull/86)).
 
+## 0.11.0 - 08-25-2026
+
 ### Changed
 
 * Event channels are now passed to `run` instead of at the time of backend
   construction ([#85](https://github.com/stjude-rust-labs/crankshaft/pull/85)).
 * `Execution`s can now specify multiple container images to act as fallbacks ([#83](https://github.com/stjude-rust-labs/crankshaft/pull/83)).
 * `Backend::run()` now returns a collection of `ExecutionResult`s, rather than `ExitStatus` ([#83](https://github.com/stjude-rust-labs/crankshaft/pull/83)).
+
+### Dependencies
+
+* Updated dependencies to latest ([#87](https://github.com/stjude-rust-labs/crankshaft/pull/87)).
 
 ## 0.10.0 - 04-22-2026